requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bidict==0.23.1
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
fastapi==0.115.4
fastapi-socketio==0.0.10
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
parsimonious==0.10.0
pluggy==1.5.0
poly_eip712_structs==0.0.1
py_clob_client==0.18.0
py_order_utils==0.3.2
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.7.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.2
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
wsproto==1.2.0
